requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose
aiosmtplib
email-validator
pydantic
cloudinary